+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features
 1. [#510](https://github.com/influxdata/influxdb-client-python/pull/510): Allow to use client's optional configs for initialization from file or environment properties
 
+### Bug Fixes
+1. [#512](https://github.com/influxdata/influxdb-client-python/pull/512): Exception propagation for asynchronous `QueryApi` [async/await]
+
 ## 1.33.0 [2022-09-29]
 
 ### Features